facial-recognition/requirements.txt
azure-core==1.31.0
azure-storage-blob==12.23.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dlib==19.24.4
Flask==3.0.3
Flask-Cors==4.0.1
flask-talisman==1.1.0
Flask-Testing==0.8.1
greenlet==3.0.3
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.10.0.82
pycparser==2.22
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.3